[refactor] move classes to anther pachages
