+++ src/main/java/chess/domain/board/File.java
@@ -1,4 +1,4 @@
-package chess.domain;
+package chess.domain.board;
 
 public enum File {
     A('a'),
