+++ src/main/java/chess/domain/piece/Piece.java
@@ -1,6 +1,5 @@
 package chess.domain.piece;
 
-import chess.domain.Color;
 import chess.domain.Direction;
 import chess.domain.board.Square;
