+++ src/main/java/chess/domain/piece/Type.java
@@ -1,4 +1,4 @@
-package chess.domain;
+package chess.domain.piece;
 
 public enum Type {
     ROOK("R", 5),
