+++ src/main/java/chess/view/OutputView.java
@@ -1,8 +1,8 @@
 package chess.view;
 
-import chess.domain.Color;
-import chess.domain.File;
-import chess.domain.Rank;
+import chess.domain.piece.Color;
+import chess.domain.board.File;
+import chess.domain.board.Rank;
 import chess.domain.board.Square;
 import chess.domain.piece.Piece;
