+++ src/test/java/chess/domain/TeamScoreTest.java
@@ -2,6 +2,7 @@ package chess.domain;
 
 import chess.domain.board.ChessBoard;
 import chess.domain.board.Square;
+import chess.domain.piece.Color;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
