+++ src/test/java/chess/domain/board/ChessBoardTest.java
@@ -1,6 +1,6 @@
 package chess.domain.board;
 
-import chess.domain.Color;
+import chess.domain.piece.Color;
 import chess.domain.piece.Pawn;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
