+++ src/test/java/chess/domain/piece/BishopTest.java
@@ -1,7 +1,6 @@
 package chess.domain.piece;
 
 
-import chess.domain.Color;
 import chess.domain.board.Square;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
